+++ packages/placeholder-pdfkit010/src/index.js
@@ -1,0 +1,3 @@
+export * from './pdfkitAddPlaceholder';
+export * from './pdfkitReferenceMock';
+export {default as PDFObject} from './pdfkit/pdfobject';