+++ packages/placeholder-plain/jest.config.js
@@ -1,0 +1,4 @@
+const sharedConfig = require('../../jest.config.base');
+module.exports = {
+    ...sharedConfig,
+};